Allow for disabling force push
If the action is pushing to the same branch, it is likely desirable to
do a regular push instead of a force which might overwrite other
changes in case the source has changed after the action triggered.

To not break the interface of the action, add a boolean flag to turn
on 'NO_FORCE_PUSH' mode.